% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annulus.R
\name{runif_in_annulus}
\alias{runif_in_annulus}
\title{Uniform sampling in an annulus}
\usage{
runif_in_annulus(n, O, r1, r2)
}
\arguments{
\item{n}{number of simulations}

\item{O}{center of the annulus}

\item{r1}{inner radius}

\item{r2}{outer radius}
}
\value{
The simulations in a \code{n} times \code{2} matrix.
}
\description{
Uniform sampling in an annulus (dimension 2).
}
\examples{
sims <- runif_in_annulus(100, c(0, 0), 1, 2)
plot(sims, xlim = c(-2, 2), ylim = c(-2, 2), asp = 1, pch = 19)
}
