% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pball.R
\name{runif_in_pball}
\alias{runif_in_pball}
\title{Uniform sampling in a p-ball}
\usage{
runif_in_pball(n, d, p, r = 1)
}
\arguments{
\item{n}{number of simulations}

\item{d}{dimension}

\item{p}{exponent in the p-norm, a positive number}

\item{r}{positive number, the radius}
}
\value{
The simulations in a \code{n} times \code{d} matrix.
}
\description{
Uniform sampling in a p-ball (arbitrary dimension).
}
\examples{
sims <- runif_in_pball(500, d = 2, p = 1)
plot(sims, xlim = c(-1, 1), ylim = c(-1, 1), asp = 1)
}
