% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{runif_triangle}
\alias{runif_triangle}
\alias{runif_in_triangle}
\alias{runif_on_triangle}
\title{Uniform sampling on/in a triangle}
\usage{
runif_in_triangle(n, v1, v2, v3)

runif_on_triangle(n, v1, v2, v3)
}
\arguments{
\item{n}{number of simulations}

\item{v1, v2, v3}{vertices of the triangle}
}
\value{
The simulations in a \code{n} times \code{2} matrix.
}
\description{
Uniform sampling on or in a triangle (dimension 2).
}
\examples{
sims <- runif_on_triangle(30, c(0,0), c(1,0), c(0,1))
plot(sims, xlim = c(0,1), ylim = c(0,1), pch = 19)
sims <- runif_in_triangle(100, c(0,0), c(1,0), c(0,1))
plot(sims, xlim = c(0,1), ylim = c(0,1), pch = 19)
}
