% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_sphere}
\alias{surface_sphere}
\title{Sphere surface}
\usage{
surface_sphere(d, r = 1)
}
\arguments{
\item{d}{dimension of the space}

\item{r}{radius of the sphere}
}
\value{
The surface of the sphere of radius \code{r} in the
\code{d}-dimensional space.
}
\description{
Surface of a sphere (arbitrary dimension).
}
\examples{
r <- 2
surface_sphere(3, r)
4*pi*r^2
# perimeter of the unit circle:
surface_sphere(2)
}
