% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_sphericalCap}
\alias{surface_sphericalCap}
\title{Spherical cap surface}
\usage{
surface_sphericalCap(r, h)
}
\arguments{
\item{r}{radius of the sphere}

\item{h}{height of the cap}
}
\value{
The surface area of the spherical cap.
}
\description{
Surface of a spherical cap.
}
