% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_stri}
\alias{surface_stri}
\title{Spherical triangle surface}
\usage{
surface_stri(r, v1, v2, v3)
}
\arguments{
\item{r}{radius}

\item{v1, v2, v3}{vertices}
}
\value{
The surface of the spherical triangle of radius \code{r} with
vertices \code{v1}, \code{v2}, \code{v3}.
}
\description{
Surface of a spherical triangle.
}
\examples{
# surface of the first orthant:
surface_stri(r=1, c(1,0,0), c(0,1,0), c(0,0,1))
}
