% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_torus}
\alias{surface_torus}
\title{Torus surface}
\usage{
surface_torus(R, r)
}
\arguments{
\item{R}{major radius}

\item{r}{minor radius}
}
\value{
The surface area of the torus.
}
\description{
Surface of a torus.
}
