% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_triangle}
\alias{surface_triangle}
\title{Triangle surface}
\usage{
surface_triangle(v1, v2, v3)
}
\arguments{
\item{v1, v2, v3}{vertices of the triangle}
}
\value{
The surface of the triangle with vertices \code{v1}, \code{v2},
\code{v3}.
}
\description{
Surface of a triangle.
}
\examples{
surface_triangle(c(0,0), c(0,1), c(1,0))
}
