% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_pball}
\alias{volume_pball}
\title{p-ball volume}
\usage{
volume_pball(d, p, r = 1)
}
\arguments{
\item{d}{dimension}

\item{p}{exponent in the p-norm, a positive number}

\item{r}{radius of the ball}
}
\value{
The volume of the \code{p}-ball with radius \code{r}.
}
\description{
Euclidean volume of a p-ball (arbitrary dimension).
}
\examples{
volume_pball(d=4, p=2, r=2)
volume_sphere(d=4, r=2)
}
