% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_sphere}
\alias{volume_sphere}
\title{Sphere volume}
\usage{
volume_sphere(d, r = 1)
}
\arguments{
\item{d}{dimension of the space}

\item{r}{radius of the sphere}
}
\value{
The volume of the sphere with radius \code{r} in the
\code{d}-dimensional space.
}
\description{
Volume of a sphere (arbitrary dimension).
}
\examples{
r <- 2
volume_sphere(3, r)
4/3*pi*r^3
}
