% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_sphericalCap}
\alias{volume_sphericalCap}
\title{Spherical cap volume}
\usage{
volume_sphericalCap(r, h)
}
\arguments{
\item{r}{radius of the sphere}

\item{h}{height of the cap}
}
\value{
The volume of the spherical cap.
}
\description{
Volume of a spherical cap.
}
