% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_tetrahedron}
\alias{volume_tetrahedron}
\title{Tetrahedron volume}
\usage{
volume_tetrahedron(v1, v2, v3, v4)
}
\arguments{
\item{v1, v2, v3, v4}{vertices of the tetrahedron}
}
\value{
The volume of the tetrahedron.
}
\description{
Volume of a tetrahedron (dimension 3).
}
\examples{
v1 <- c(0,0,0); v2 <- c(1,0,0); v3 <- c(0,1,0); v4 <- c(0,0,1)
volume_tetrahedron(v1, v2, v3, v4)
volume_unitSimplex(3)
}
\seealso{
\code{\link{volume_simplex}} for the volume of a simplex in
arbitrary dimension.
}
