% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_torus}
\alias{volume_torus}
\title{Torus volume}
\usage{
volume_torus(R, r)
}
\arguments{
\item{R}{major radius}

\item{r}{minor radius}
}
\value{
The volume of the torus.
}
\description{
Volume of a torus.
}
