% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes.R
\name{volume_unitSimplex}
\alias{volume_unitSimplex}
\title{Unit simplex volume}
\usage{
volume_unitSimplex(d)
}
\arguments{
\item{d}{dimension of the space}
}
\value{
The volume of the unit simplex in the space of dimension \code{d}.
}
\description{
Volume of the unit simplex (arbitrary dimension).
}
\seealso{
\code{\link{volume_simplex}} for the volume of an arbitrary simplex.
}
